New: DiatonicInterval class.
The class implements in the pitchtools module.

The class is a value object whose values can not be changed after
instantiation.

Abjad diatonic intervals have a basic understanding of their actual
value in semitones.

The class is basically a stub class at the moment. Will be used
for a diatonic transposition function later. Will also be used
for calculating intervals of string harmonics (also later).

Examples of repr:

   abjad> pitchtools.DiatonicInterval('perfect', 1)
   DiatonicInterval(perfect unison)
   abjad> pitchtools.DiatonicInterval('minor', 2)
   DiatonicInterval(ascending minor second)
   abjad> pitchtools.DiatonicInterval('major', 2)
   DiatonicInterval(ascending major second)

Examples of semitone calculation:

   abjad> pitchtools.DiatonicInterval('perfect', 1).semitones
   0
   abjad> pitchtools.DiatonicInterval('minor', 2).semitones
   1
   abjad> pitchtools.DiatonicInterval('major', 2).semitones
   2