Deprecation: previous commits implemented three receipt class in abjad/receipt:
   1. _ComponentReceipt
   2. _SlipReceipt
   3. _SpannerReceipt

Of these, numbers 1 and 2 weren't being used at all. Number 3, _SpannerReceipt,
was being used in the implementation of _Component._detach( ) and 
_Component._reattach( ). However, two things are important here.
The first is that _Component._reattach( ) was never being used.
The second is that _Component._detach( ) doesn't need to return a
_SpannerReceipt. In fact, _Component._detach( ) doesn't need to return
any value at all.

All of this means that all three receipt classes are now completley deprecated.
This commit comments out all three receipt classes, as well as the definition
of _Component._reattach( ) and the return value of _Component._detach( ).

The next commit will remove this commented-out code.