New:
   * tuplettools.augmentation_to_diminution(tuplet)
   * tuplettools.diminution_to_augmentation(tuplet)

The first function multiplies the written duration of the leaves in tuplet
by the least power of 2 necessary to diminish an augmented tuplet.

The second function divides the written duration of the leaves in tuplet
by the least power of 2 necessary to augment a diminshed tuplet.

That is, these functions double, or halve, leaf durations in tuplet,
respectively.

Example:

      abjad> tuplet = FixedDurationTuplet((2, 4), construct.scale(3))
      abjad> tuplet
      FixedDurationTuplet(1/2, [c'8, d'8, e'8])
      abjad> tuplettools.augmentation_to_diminution(tuplet)
      FixedDurationTuplet(1/2, [c'4, d'4, e'4]) 

Example:

      abjad> tuplet = FixedDurationTuplet((2, 8), construct.scale(3))
      abjad> tuplet
      FixedDurationTuplet(1/4, [c'8, d'8, e'8])
      abjad> tuplettools.diminution_to_augmentation(tuplet)
      FixedDurationTuplet(1/4, [c'16, d'16, e'16])