Changed:
   scoretools.find(t) now returns first component found.
   scoretools.find(t) previously returned tuple of all components found.
   
This commit breaks all calls to scoretools.find(t).
You should change

   violin = scoretools.find(t, 'Violin')[0]

to

   violin = scoretools.find(t, 'Violin')

In the case that no component is found, scoretools.find( ) now raises
the new MissingComponentError:

   scoretools.find(t, 'foo')
   MissingComponentError

It's annoying that this change will break all calls to scoretools.find(t)
without exception. But the new behavior provided by the function is
much better than before.

If you want to iterate over all components matching a name, class or
context, use the extensive set of functions in the iterate module.