New:
   * durtools.prolated_to_prolation_written_pairs(prolated_duration)

The function exists to answer the question "what are the different
ways to notation prolated_duration as some note under some tuplet
bracket?"

Example:

   abjad> pairs = durtools.prolated_to_prolation_written_pairs(Rational(1, 12))
   abjad> for pair in pairs: pair
   ... 
   (Rational(2, 3), Rational(1, 8))
   (Rational(4, 3), Rational(1, 16))
   (Rational(8, 9), Rational(3, 32))
   (Rational(16, 9), Rational(3, 64))
   (Rational(16, 21), Rational(7, 64))
   (Rational(32, 21), Rational(7, 128))
   (Rational(32, 45), Rational(15, 128))

The output above gives seven different ways of notation a duration of 1/12.

The first option is (Rational(2, 3), Rational(1, 8)) which means
"a written 1/8th note under a 3:2 bracket." This is the most usual
option.

The second option is (Rational(4, 3), Rational(1, 16)) which means
"a written 1/16th note under a 3:4 bracket."

The third ooption is (Rational(8, 9), Rational(3, 32)) which means
"a written dotted 1/16th note under a 9:8 bracket."

And so on.

The optional minimum_written_duration keyword is set to a 1/128th note.
You can set this value to 1/256 or 1/512, etc, to generate considerably
more exotic possibilities.