Updated demo "desordre". The demo broke due to some of the several changes that have been taking place. The demo now works again. This update is the first step into moving this demos to the "Examples" section in the documentation.
A general question that here arises is, given that we have the documentation examples, does it make sense to still have the "demos" code? Are they still useful? 

Either way, neither the demos nor the documentation "examples" are tested when we run py.test, which is why these every now and then go out of sync without out noticing. We definitely don't want the documentation examples to be broken, so we should find a way to run the code of the documentation examples via the py.test.