New:
   * persistencetools.pitch_and_rhythm_skeleton( )

Skeletonization has returned.

Reduced, bare bones alternative to the built-in pickle and cPickle modules.

Note that the function preserves ONLY pitches and rhythms; almost
everything else is lost.

Stuff that gets lost includes:

   * clefs
   * key signatures
   * overrides
   * etc

Essentially, all ATTRIBUTES ARE LOST but all PITCH AND RHYTHMIC STRUCTURE
is preserved.

The idea will be to NOT extend this function later but instead to keep
it as a minimal, bare bones alternative to pickle and cPickle.

The idea here is to use this function very early in the compositional
process when you may want a fast, lightweight way to save posibly
large amounts of basic pitch or rhythmic information, but before
you have started to set the visual attributes of the score itself.