Extension:
   * listtools.phasor( )

The function now allows generator input instead of only list input.

Also, the function now performs exactly twice as fast after optimization
using itertools.cycle( ) and itertools.islice( ).


New:

   * listtools._generator(*args)

Private function that is not part of the public API but is useful
to easily create a generator with a single line of code for test cases.

Example:

   abjad> listtools._generator(1, 8)
   <generator object at 0x6b148>
   abjad> for x in _:
   ...     x
   ... 
   1
   2
   3
   4
   5
   6
   7