Extension: PitchClassSet now inherits from built-in frozenset instead of from built-in set.
This means that PitchClassSet instances are now hashable and can now
be contained in built-in set instances.

This also means that PitchClassSet instances, like other frozenset instances,
no longer implement add( ) or update( ) methods. That is, PitchClassSet
content is now immutable. This is a good thing since PitchClassSet
instances should be treated as value objects.