New:
   lilytools.parse_note_entry_string( )

Parse simple LilyPond note entry string.

Example:

      abjad> note_entry_string = "g'2 a'2 g'4. fs'8 e'4 d'4"
      abjad> lilytools.parse_note_entry_string(note_entry_string)
      [Note(g', 2), Note(a', 2), Note(g', 4.), Note(fs', 8), Note(e', 4), Note(d', 4)]


NOTE: The parse is primitive. Only notes, rests and skips are recognized.
Tuplets are not recognized. Time signature, key signature and other directives
are not recognized.

The purpose of the function is to facilitate the entry of tonal music into
Abjad for analytic applications. The function works, for example,
to quickly entry the voices of a Bach chorale.