New:
   tools.tonalharmony

The module exists to test, label and create objects from tonal harmony.

Current contents:

   tonalharmony.is_unlikely_melodic_diatonic_interval_in_chorale( )
   tonalharmony.mark_unlikely_melodic_intervals_in_chorale( )

We can expect the module to grow considerably, depending on the number 
of different styles of tonal music to be analyzed.

Initial check-ins will include functions for the analysis of
the Bach chorales, ie, tertian harmony up to seventh chords,
phrase and cadence analysis, standard voice-leading rules in the style
of the chorales, roman numeral anlaysis, etc.