Modification for port to Python 2.6.
Python 2.6 introduces a public 'format' attribute on the built-in string type.
This 'format' attribute was not present in Python 2.5 and earlier.

The Abjad _Parser class implemented in core/parser.py now implements a
two-line check to handle built-in strings directly rather than looking
for a 'format' attribute, as had been the case up until this release.

All tests pass under 2.5; tests not yet run explicitly under 2.6.