The Sphinx makefile is located here:
   ABJADPATH/documentation/Makefile

The Sphinx makefile includes targets for doctest, html, latex
and other output formats.

This commit adds a new target called 'allhtml' to the Sphinx makefile.

The purpose of the new 'allhtml' target in the Sphinx makefile is
to allow the regeneration of the entire doc tree rather than
just regenerating those files in the doc tree that have recently changed.

Usage:

   1. Change to ABJADPATH/documentation
   2. Then call

      make allhtml

That's all.