Sphinx 1.0b2 comes with nine themes:
   * agogo
   * basic
   * default
   * epub
   * haiku
   * nature
   * scrolls
   * sphinxdoc
   * traditional

Of these nine built-in Sphinx themes, the Haiku theme
solves line-overflow errors in the Abjad API.

This commit sets the Abjad docs to render with the Sphinx Haiku theme.

Note the Haiku theme is not the perfect long-term solution for the
Abjad docs because the Haiku theme uses none of the Abjad colors.

The Haiku theme is also missing a few layout features relative
to the custom Abjadtheme theme.

But this commit at least temporariliy fixes 
line-overflow problems in the API.

Note also that the non-API pages of the docs look worse with the Haiku
theme. Even thought the API pages of the docs look better.