Renamed 49 test files because of a bug in the new version of py.test.
The py.test bug appears in the following circumstances:

Test files with names in the form foo___init__.py cause py.test to blow up.

The filename must end in ___init__.py exactly, which, bizarrely,
was precisely the case for 49 Abjad test files designed to test the
__init__( ) method of custom classes in the system.

All such 49 test files are renamed in this commit to
the form foo___init___.py which insert one additional underscore _
prior to the .py extension.

All tests now pass with the new version of py.test.