New tools.macros package.
The package exists to define a limited number of system macros.

Current macro definition:

   macros.scale ==
   leaftools.make_first_n_notes_in_ascending_diatonic_scale

The tools.macros package should define only very macros
because macros decrease the overall readability of the source.

The justification for the definition of macros.scale( ) is that
the function is used in almost every test file and every doc
example in the codebase.

Note that the tools.macros package is not yet documented in the API.
The tools.macros package will eventually be documented on the doc
site at a location separate from that of the other tools packages.