New abjad/optimization directory. Current contents:
   r3458
   r3458+Fraction

The purpose of the directory is to house optimization-related findings.

These first two files baseline Note creation and formatting for r3458
and then repeat the same tests with the Python 2.6 Fraction class
replacing the Abjad Rational class. The conclusion of the tests
shown in r3458+Fraction indicate no speed increase with Fraction
but a memory reduction of between 9 - 10 %, that is from
84.8M (r3458) to 82.3M (r3458 + Fraction optimization) for the memory
consumption of 1000 notes as indicated by the heapy module.

This commit checks in Fraction as a permanent replacement from Rational.

Python 2.6 now required from this commit forward.