Entirely new pattern: OverrideNamespace / GrobNamespaces.
It is now possible to specify LilyPond grob overrides in a new way:

   abjad> note = Note(0, (1, 4))
   abjad> note.override.accidental.color = 'red'
   abjad> note.override.beam.positions = (-6, -6)
   abjad> note.override.dots.thickness = 2

   abjad> f(note)
   \once \override Accidental #'color = #red
   \once \override Beam #'positions = #'(-6 . -6)
   \once \override Dots #'thickness = #2
   c'4

In this example, note.override references the new OverrideNamespace.
The new OverrideNamespace initializes empty with no attributes of its own.
As the the composer dynamically adds attributes to the OverrideNamespace,
the OverrideNamespace instantiates new (and very thin) GrobNamespace objects.
The attribute note.override.accidental references a new GrobNamespace object.
As the composer dynamically adds attributes to GrobNamespace objects,
the GrobNamespace objects collect those attributes in anticipation of format-time.
At format-time, the existing InterfaceAggregator (which has been in the system
for some time now) of each Abjad component collects LilyPond override statements 
from the new OverrideNamespace that now attaches to each Abjad component.
In turn, the new OverrideNamespace collects LilyPond override statements from
each of the (zero to many) GrobNamespace objects that the composer has dynamically
added to the OverrideNamespace during the work of composition.
LilyPond override statements are then output at format-time as before.

The motivation here is that the collection of many dozens of interfaces
attaching to each Abjad component looks to be the primary consumer of memory
resources globally across the entire system. Every interface removed from the core
_Component class looks to reduce function calls at construction, to reduce
function calls at format-time, and to reduce overall system memory.
This is the reason for the new OverrideNamespace / GrobNamespace pattern
introduced in this commit.

Note that this commit leaves absolutely all previous functionality of the system
in tact; all tests complete without modification. But note that this is likely
to change in the very near future as many of the existing dozens of interfaces
are deprecated. Note that this will force composes to use the new
component.override.grob.attribute = value syntax in place of the previous
component.grob.attribute = value formulation.