New pattern: arbitrary LilyPond context settings.
Example:

   abjad> staff = Staff(macros.scale(4))
   abjad> staff.set.suggest_accidentals = False
   abjad> f(staff)
   \new Staff \with {
      suggestAccidentals = ##f
   } {
      c'8
      d'8
      e'8
      f'8
   }

Example:

   abjad> staff = Staff(macros.scale(4))
   abjad> staff[1].set.suggest_accidentals = False
   abjad> f(staff)
   \new Staff {
      c'8
      \set suggestAccidentals = ##f
      d'8
      e'8
      f'8
   }

The examples work because of the new LilyPondContextSettingNamespace 
now implemented as the read-only 'set' attribute on all Abjad components.

All LilyPond context settings may be made with the new 'set' attribute
as of this commit, though it should be noted that so-called "promoted"
context settings for the form \set Staff.suggestAccidental = ##f will
follow in a later commit.

The motivation behind the new LilyPond context setting namespace is
to allow the deprecation of even more _Interface-derived classes in
the near future to reduce the in-memory footprint of score components.