Entirely new pattern: LilyPondMiscellaneousCommandComponentPlugIn.
Example:

   abjad> staff = Staff(macros.scale(4))
   abjad> staff.misc
   LilyPondMiscellaneousCommandComponentPlugIn( )

Example:

   abjad> staff.misc.set_accidental_style = 'forget'
   abjad> f(staff)
   \new Staff {
      #(set-accidental-style 'forget)
      c'8
      d'8
      e'8
      f'8
   }

The purpose of the class is to afford composes a Pythonic way of
issuing LilyPond commands that -- for whatever reason internal to LilyPond --
take the form of neither grob overrides nor context settings.

The class implements a private, groomed list of these commands;
arbitrary assignments to component.misc raise an attribute error.

This 'miscellaneous command' pattern means that even more _Interfaces
can be removed. This commit removes AccidentalInterface as proof of concept.