Renamed core/_Abjad to core/_StrictComparator.
The previous name was misleading because not all system objects
inherit from this class. Only this system objects needing to implement
Python 3.0-style strict comparison inherit from this class. Of course
this is still a large number of system classes.

The new _StrictComparator name harmonizes with _Immutable
and _ImmutableDictionary. All are abstract base classes designed
to mix in to custom classes in the cases where certain high-level
behaviors are desired.