Renamed testfile with '__init__' in filename:
   test_NumberedChromaticPitchClassVector___init__.py
   ==>
   test_NumberedChromaticPitchClassVector___init___.py

Py.test introduced a bug in the 1.3.2 py.test release.

The bug causes py.test to error on testfiles with '__init__' in filename.

The bug persists in py.test releases 1.3.3 and 1.3.4.

Until the py.test bug is removed in a future py.test release.
the workaround continues to be to name testfiles as shown above,
avoid the exact string '__init__'.