Extended leaves with two new public attributes:
   * _Leaf.pitch_indication_is_at_sounding_pitch
   * _Leaf.pitch_indication_is_nonsemantic

Both are read / write boolean flags.

The uses of the attributes are as follows:

_Leaf.pitch_indication_is_at_sounding_pitch is set to true by default. The attribute
is to stay set to true until such time as the pitch content of the leaf is transposed.
This happens, for example, when score in C is transposed up or down for the parts to
be given to the player of a transposing instrument. Once leaves are transposed this
attribute is to remain set to false. So you can think of this attribute as a boolean
'is_at_concert_pitch' flag, or, equivalently, as a boolean 'is_not_yet_transposed' flag.

_Leaf.pitch_indication_is_nonsemantic is set to false by default. The attribute is stay
set to false until such as the composer chooses to indicate that the leaf in question no
longer indicates pitch in any real way. This happens when, for example, leaves are used to 
populate a tab special tab staff that indicates (possibly parametric) physical gestures
rather than pitches. This is also happens when leaves are used to represent unpitched percussion.
Leaves with this attribute set to true will eventually be left out of almost all pitch 
operations and analyses. Note that setting this attribute to true automatically sets
_Leaf.pitch_indicaiton_is_at_sounding_pitch (described above) to false because the distinction
between sounding pitch / transposed pitch is not relevant for notes the pitch content of which
is taken to be nonsemantic anyway.

Note that rests current implement both of these public attributes. That will probably change.