Removed small number of UNICODE characters from docs to work around bug in Sphinx LaTeX builder.
Summary: UNICODE characters are *mostly* fine when the Sphinx LaTeX builder runs. But there's a known
bug identified on the Sphinx dev site that the LaTeX builder grumbles when 'ids' contain UNICODE characters.
Turns out that ids are things like [Adán2006] in the bibliography and even the acute-a causes problems.
Fortunately there are not many of those characters and so they have simply been removed in this commit.

More: removals described above make the Sphinx LaTeX builder run. The resulting .../docs/_build/latex/Abjad.tex
file still has to be run through pdflatex with the command 'make all-pdf'. The remaining changes in this
commit remove further UNICODE characters that garble 'make all-pdf'. These included the set-membership epsilon
in the API entry to one function and the sharp and flat signs in the appendix on pitch conventions.

LaTeX docs build correctly as of this commit.