Fixed Google #64.
Background: in previous commits the following would occur:

   staff = Staff([ ])
   contexttools.get_effective_time_signature(staff)
   time_signature_mark = contexttools.TimeSignatureMark(3, 4)
   time_signature_mark(staff)
   f(staff)
   \new Staff {
      %%% \time 3/4 %%%
   }

The triple percent signs indicate a failure because no effective time signature
can be determined for the staff. (That is, at format time there is a check to find
an effective time signature for the staff and none is found. The reason the format-time
code knows about the time signature at all is because the 'marks' attributing attaching
to staff contains the time signature mark; but the time signature mark has not yet
been appropriately updated to find its effective context. Abjad takes this as evidence
of a broken mark and outputs the triple percents accordingly.)

This turned out to be the case because the context mark attachment code was being broken
by a subtle bug. First, ContextMark inherits from Mark. Second, ContextMark.attach_mark( ) 
is the same as Mark.attach_mark( ). Third, Mark.attach_mark( ) contained a call to the mangled
method Mark.__bind_start_component( ). Fourth, ContextMark.attach_mark( ) was therefore
calling Mark.__bind_start_component( ) instead of ContextMark.__bind_start_component( );
and this was the point of misunderstanding: mangled methods (beginning in double underscore)
can not only not be called by alien classes but can not, apparently, even be called
*by inheriting classes*.

The fix was then the following renames:

   Mark.__bind_start_component( )
   Mark._bind_start_component( )

   ContextMark.__bind_start_component( )
   ContextMark._bind_start_component( )