In the tools packages: replace many Fraction( ) calls with calls to Duration( ).
Background: the Abjad Duration class is nothing more than a slightly smarter synonym for 
Python's built-in Fraction. The difference is that Duration has a more forgiving initialization
signature than Fraction; specifically Duration((1, 2)) works where Fraction((1, 2)) doesn't.
This matters because it enbables the global use of (1, 2)-style 'duration tokens' across
the Abjad codebase.

More background: the commit here touches 323 files. All 323 files are in the various tools packages.
Most (but not all) of the calls to Fraction( ) in the tools packages are replaced with calls to 
Duration( ) in this commit. The change wasn't a blind search-and-replace because the difference
between durations and fractions is semantic: durations properly model objects consuming musical
time; fractions properly model things like multipliers.