Removed private Measure._explicit_meter attribute.
The previous regime of the _explicit_meter attribute conflicted with the
more general principles of context marks and the way that they attach to
score components.

Measures are now simply components to which you can attach and detach
a time signature mark. Just as before, at measure-initialization time
a time signature mark is created and attached to the measure; but this
is now the only special relationship between measures and time signature marks.