Removed comments from previous commits involved in removing Measure._explicit_meter.
Fixed broken doc tests in quantizationtools.

Note that doctest complains about docstrings formatted like this:

class QGridDeletionTree(_Immutable):
   '''A model of the nested division structure of a Q-grid,
   and of which timepoints have been deleted in that grid.

   ::

      >>> tree.format_for_beatspan(Fraction(1, 4))
      Tuplet(2/3, [c'8, {c'16, r16}, {* 3:2 c'16, r16, r16 *}])
   '''

Why?

The above is perfectly valid ReST. (And doesn't, for example,
cause Sphinx to complain.) But doctest raises a ValueError like this:

< many lines of traceback omitted >
  File "/System/Library/Frameworks/Python.framework/Versions/2.6/lib/python2.6/doctest.py", line 729, in _check_prefix
    (lineno+i+1, name, line))
ValueError: line 38 of the docstring for QGridDeletionTree.py has inconsistent leading whitespace: "   '''"

Turns out that what this means is that doctest is unhappy with *the last line immediately
before the close-triple-quote* not having text that aligns *exactly with the beginning
of the close-triple-quote*.

So the fix is to include a final dummy-sentence of text in the docstring immediately
before the close-triple-quote:

class QGridDeletionTree(_Immutable):
   '''A model of the nested division structure of a Q-grid,
   and of which timepoints have been deleted in that grid.

   ::

      >>> tree.format_for_beatspan(Fraction(1, 4))
      Tuplet(2/3, [c'8, {c'16, r16}, {* 3:2 c'16, r16, r16 *}])

   Return newly constructed QGridDeletionTree.
   '''

Then doctest no longer complains.

Note that this is true for all several thousand doc strings in the system.
So it's good to be in the habit of always checking the conclusion of each
new docstring written to make sure it adheres to this format.