Extended _Component with two private methods:
   * _Component._forbid_component_to_update( )
   * _Component._allow_component_to_update( )

The methods are to be used by developers only.

The methods are to be used in a pair in the implementation
of complex tools functions that iterate through an arbitrary
expression and make structural changes along the way:

   component._update_marks_of_entire_score_tree_if_necessary( )
   component._update_prolated_offset_values_of_entire_score_tree_if_necessary( )
   try:
      component._forbid_component_to_update
      < shielded structural changes go here >
   finally:
      component._allow_component_to_update( )

See measuretools.fuse_contiguous_measures_in_container_cyclically_by_counts( )
for an example.