Removed _*FormatterNumberInterface classes.
This commit removes the old regime of 'number formatters'.

Background: in ancient versions of the system, the formatter number 
interface classes afforded the labeling of leaves with leaf indices and
the numbering of measures with measure numbers. Leaf indices and measure
numbers could be labeled either with markup or else with comments. But
the pattern was unusual in the system and so the code tended to be 
maintained only intermittently.

As of this commit all formatter number interface classe have been removed. 
Now you should use markup to label
leaves with leaf index markup and you should use comments to label
leaves with leaf index comments. Likewise, you should use markup to label
measures with measure number markup and you should use comments to label
measures with measure number comments.