Added mark tools.
The API now features to parallel sets of getter functions.

Functions named in the singular:

    marktools.get_annotation_attached_to_component
    marktools.get_articulation_attached_to_component
    marktools.get_lilypond_command_mark_attached_to_component
    marktools.get_lilypond_comment_attached_to_component
    marktools.get_mark_attached_to_component
    marktools.get_noncontext_mark_attached_to_component
    marktools.get_stem_tremolo_attached_to_component

Functions named in the plural:

    marktools.get_annotations_attached_to_component
    marktools.get_articulations_attached_to_component
    marktools.get_lilypond_command_marks_attached_to_component
    marktools.get_lilypond_comments_attached_to_component
    marktools.get_marks_attached_to_component
    marktools.get_noncontext_marks_attached_to_component
    marktools.get_stem_tremolos_attached_to_component

The difference is as follows:

Functions named in the singular return exactly one articulation (or other object)
if there is articulation to return. Otherwise, such functions raise MissingMarkError
(if no mark is attached to the component on which the function is called) or raise
ExtraMarkError (if more than one such mark is attached to the component on which
the function is called).

Functions named in the plural return a tuple of zero or more marks (of the appropriate kind)
attached to the component on which the function is called.

So the functions named in the singular are strict; functions named in the plural
as lenient.