Filled in the public API for context mark getter functions.
There is now a getter function in the singular for each context mark:

    contexttools.get_clef_mark_attached_to_component
    contexttools.get_context_mark_attached_to_component
    contexttools.get_dynamic_mark_attached_to_component
    contexttools.get_instrument_mark_attached_to_component
    contexttools.get_key_signature_mark_attached_to_component
    contexttools.get_staff_change_mark_attached_to_component
    contexttools.get_tempo_mark_attached_to_component
    contexttools.get_time_signature_mark_attached_to_component

There is also a getter function the plural for each context mark:

    contexttools.get_clef_marks_attached_to_component
    contexttools.get_context_marks_attached_to_component
    contexttools.get_dynamic_marks_attached_to_component
    contexttools.get_instrument_marks_attached_to_component
    contexttools.get_key_signature_marks_attached_to_component
    contexttools.get_staff_change_marks_attached_to_component
    contexttools.get_tempo_marks_attached_to_component
    contexttools.get_time_signature_marks_attached_to_component

The difference is as follows:

Call a getter function in the singular when you are certain that *exactly one*
context mark (of the appropriate type) attaches to the component in question.
If exactly one context mark does indeed attach to the component in question,
then the function returns that context mark. If no context mark (of the 
appropriate type) attaches to the component in question then the function
raises a MissingMarkError.

Call a getter function in the plural when you want to return however many
context marks (of the appropriate type) attach to the component in question.
All such plural getters return a tuple of zero or more context marks (of the
appropriate type). Plural getters raise no exceptions.