Extended Spanner with trivial __lt__ implementation based on repr.
Reason: to allow canonic display order of spanners attaching to a component.

(The spanners attaching to a component are kept in an unordered set because 
the attachment order of spanners relative to a component makes no semantic sense.
This works fine except that there is no canonic way to display the elements of
an unordered set, save by sorting the elements. It is now possible to sort
an unordered set of spanners and, hence, make for happier doctests.)