Added 'require' decorator.
The decorator forces argument type checking and works like this:

    from abjad.tools.iotools.require import require

    @require(int)
    def foo(x):
        return x ** 2

This is basically the same as saying:

    def foo(x):
        assert isinstance(x, int)
        return x ** 2

The decorator has the advantage of *externalizing* type checking from function definition.

This means that it will eventually be possible to *programmatically* determine
the argument types of every public function and method in the API.

This will require a lot of work. Specifically, the decorator will have to be added to
every public function and method in the API.

But this will also externalize all type checking from every public function and method
in the API.

As of this commit the 'require' decorator implemented as private '_require'
and should be considered experimental.

The reason for this is that Sphinx does not yet play well with the 'require' decorator.

Research remains to be done to see if Sphinx can be made to find the docstrings
attaching to require-decorated functions. If a solution can be found then the decorator
will be made public and an effort will be made over the medium-term to apply the decorator
to public functions and methods throughout the API.

OTOH, if no Sphinx solution can be found then the decorator will be abandoned.

Hopefully a Sphinx solution can be found.