Extended 'require' decorator to work with Sphinx.
The solution turned out to be a single line in the implementation of the decorator:

    wrapper.__doc__ = func.__doc__

The decorator now works correctly with Sphinx. 

The decorator is now public 'require' (instead of private '_require').

But the decorator should still be considered at least somewhat experimental.

Work remains to be done to ensure that that input types passed to the decorator
can in fact be retrieved for later use. Without this the decorator fails to
achieve 100% of what it should do. But the decorator name is public as of this
commit because it seems likely that retrieving the decorator input types
will be possible.