Documented 'require' decorator for public use.
Decorator now documented in the IO tools API pages.

Example given there:

    abjad> from abjad.tools import mathtools
    abjad> from abjad.tools.iotools.require import require

    abjad> @require(mathtools.is_nonnegative_integer, str)
    abjad> def multiply_string(n, string): return n * string

    abjad> multiply_string(2, 'bar')
    'barbar'

    abjad> multiply_string(2.5, 'bar')
    ...
    AssertionError: is_nonnegative_integer(2.5) does not return true.

Decorator target is available like this:

    abjad> multiply_string.func_closure[1].cell_contents
    <function multiply_string at 0x104e512a8>

Decorator tests are available like this:

    abjad> multiply_string.func_closure[0].cell_contents
    (<function is_nonnegative_integer at 0x104725d70>, <type 'str'>)

Return decorated function in the form of function wrapper.

This is a powerful new pattern that can be used both to test input
and to document input.

It may be worth considering a 'returns' decorator to implement comparable
functionality for the return value of public functions and methods.

If this is so then implementation and testing of such a 'returns' decorator
should happen before applying the 'require' decorator to large portions of the API.
The reason for this is that the application of 'require' and 'returns' together
will likely necessitate (slight) changes to the implementation of 'require';
better to test both decorators in combination before applying either to large
sections of the codebase.

Because 'return' is a reserved word, we will have to use 'returns'.
It will then make sense to change 'require' to 'requires'.