Implemented new PitchRangeInventory class.
The class acts as an ordered list of PitchRange objects.

The purpose of the class is to model something like palettes of different pitches
available in all part of a score.

Example:

    abjad> pitchtools.PitchRangeInventory(['[C3, C6]', '[C4, C6]'])
    PitchRangeInventory([PitchRange('[C3, C6]'), PitchRange('[C4, C6]')])

The class inherits from list.

The class overloads append, extend and __contains__ to allow not just explicit
pitch range objects but also pitch range tokens to support a more convenient interface.