Replaced one line of code in Chord.py ...
    del(self[:])

... with ...

    self.__delitem__(slice(0, len(self)))

... to make pychecker happy.

The two statements are exactly equivalent.

So it looks like it may be *the syntax* of slice assignment that causes
pychecker to error. 

(This is good because if it were *the functionality* of slice assignment
then there would be no workaround.)

So it looks like rewriting slice assignments as explicit calls to
__getitem__, __delitem__ and __setitem__ may function to work
around pychecker's bug for as long as required.