Extended the following classes with private attributes:
    * pitchtools.PitchRangeInventory
    * pitchtools.PitchRange

The classes all now implement a read-only '_repr_with_tools_package' attribute.

Examples:

    abjad> pitch_range
    PitchRange('[A0, C8]')
    abjad> pitch_range._repr_with_tools_package
    "pitchtools.PitchRange('[A0, C8]')"

    abjad> inventory
    PitchRangeInventory([PitchRange('[A0, C8]'), PitchRange('[C4, D5]')])
    abjad> inventory._repr_with_tools_package
    "pitchtools.PitchRangeInventory([pitchtools.PitchRange('[A0, C8]'), pitchtools.PitchRange('[C4, D5]')])"

The purpose of the functions is to allow the objects to be written to disk for storage.

Note: this repr-with-tools-package pattern will pervade the codebase;
      the changes committed here can be seen as an update of the codebase
      on an installment basis.

Note, too, that no user functionality in Abjad will change with the addition of this pattern.
As mentioned above, the purpose of the pattern is to allow Abjad objects to be written to disk
within files that contain a minimal number of import statements.

That is, storage files containing objects with tools-package-equipped reprs can get away
with a single import statement like 'from abjad.tools import pitchtools' rather than
having to contain, for example, 'from abjad.tools.PitchRange.PitchRange import PitchRange'
together with 'from abjad.tools.PitchRangeInventory.PitchRangeInventory import PitchRangeInventory'
together with possibly additional import statements to the tune of one import per nested
object type contained within composite objects.

This will be useful for object storage and retrieval systems built to extend Abjad.