Added new pitchtools.OctaveTranspositionMapping class.
The class is only a stub as of this commit.

The class will eventually model the input passed to this function:

    * pitchtools.transpose_chromatic_pitch_number_by_octave_transposition_mapping

The input to that function currently comprises a structured collection of lists, pairs and integers.
All of these values may be better modelled by structured collections of pitchtools objects.