Added two new core classes:
    * core._ImmutableAbjadObject
    * core._MutableAbjadObject

The pair of classes are meant to be the base objects for the entire system.

All Abjad classes will eventually inherit from one of these two classes
instead of from Python's built-in 'object' class.

The purpose of the classes is to make certainly functionality available globally
throughout the system. Examples include self.class_name as a synonym for self.__class__.__name__.

The two classes committed here differ only in the implementation of __slots__.

As of this commit the following classes all now inherit from _ImmutableAbjadObject or _MutableAbjadObject:

    _FlexEqualityComparator
    _Immutable
    _ImmutableDictionary
    _LilyPondComponentPlugIn
    _LilyPondObjectProxy
    _StrictComparator
    _UnaryComparator

This means that the vast majority of system objects all now inherit from one of these two core classes.
(And all tests pass.)

Note that a single core _AbjadObject class looks to be an impossibility due to the fact that
some system objects implement __slots__ while others do not.