Added tools/datastructuretools package.
The package will hold data structures to be reused throughout the system.

The status of the package will be as follows: public but manually loading (like mathools).

NOTE: when moving classes into the new datastructuretools package, it will make
sense to make the classes *public* rather than private. This will, for example,
allow _RedBlackTree to become RedBlackTree.