Implemented two abstract base classes in `timeintervaltools`:
    * `_TimeIntervalMixin`
    * `_TimeIntervalAggregateMixin` (inherits from `_TimeIntervalMixin`)

`TimeInterval` now inherits from `_TimeIntervalMixin` and `MutableMapping`,
rather than dict.

The purpose of this commit (and later `timeintervaltools` commits) is to reduce
duplication between classes, enforce a common public interface, and ultimately
allow aggregates of `TimeIntervals` (like `TimeIntervalTree` and the soon to be
added `TimeIntervalTreeDictionary`) to act as `TimeIntervals` themselves,
deeping the package's composite pattern.